Now its compilable, hooray!!!

+++ Flsurf/Application/Staff/UseCases/DeleteComment.cs
@@ -1,18 +1,21 @@
 ﻿using Flsurf.Application.Common.Interfaces;
 using Flsurf.Application.Common.UseCases;
+using Flsurf.Application.Staff.Perms;
 using Flsurf.Domain.Staff.Entities;
 using Flsurf.Domain.User.Enums;
+using Flsurf.Infrastructure.Adapters.Permissions;
+using Microsoft.EntityFrameworkCore;
 
 namespace Flsurf.Application.Staff.UseCases
 {
     public class DeleteComment : BaseUseCase<Guid, TicketCommentEntity>
     {
         private readonly IApplicationDbContext _context;
-        private readonly IAccessPolicy _accessPolicy;
+        private readonly IPermissionService _permService;
 
-        public DeleteComment(IApplicationDbContext dbContext, IAccessPolicy accessPolicy)
+        public DeleteComment(IApplicationDbContext dbContext, IPermissionService permService)
         {
-            _accessPolicy = accessPolicy;
+            _permService = permService;
             _context = dbContext;
         }
 
@@ -20,11 +23,19 @@ namespace Flsurf.Application.Staff.UseCases
         {
             Guard.Against.Null(commentId, nameof(commentId));
 
-            var comment = await _context.TicketComments.FindAsync(commentId);
+            var comment = await _context.TicketComments
+                .Include(x => x.CreatedBy)
+                .Include(x => x.Files)
+                .FirstOrDefaultAsync(x => x.Id == commentId);
 
             Guard.Against.Null(comment, $"Comment with ID {commentId} does not exist.");
 
-            await _accessPolicy.EnforceIsAllowed(PermissionEnum.delete, comment);
+            var owner = await _permService.GetCurrentUser(); 
+
+            if (owner.Id != comment.CreatedBy.Id)
+            {
+                throw new AccessDenied("you are not owner"); 
+            }
 
             _context.TicketComments.Remove(comment);
             await _context.SaveChangesAsync();
